requirements.txt
facebook-business==24.0.1
pandas==2.2.2
python-dotenv==1.0.1
openpyxl==3.1.2
streamlit==1.35.0
plotly==5.22.0
requests==2.31.0
google-genai
